Move get_media and save_file functions to separate file
